公開 API は非同期 (web worker) 版に統一する
