+++ examples/lyra-benchmark.html
@@ -52,8 +52,8 @@
           const decoders = {};
           for (const [version, wasmUrl] of Object.entries(WASM_URL_MAP)) {
               const lyraModule = await Shiguredo.LyraModule.load(wasmUrl, MODEL_URL);
-              encoders[version] = lyraModule.createEncoder();
-              decoders[version] = lyraModule.createDecoder();
+              encoders[version] = await lyraModule.createEncoder();
+              decoders[version] = await lyraModule.createDecoder();
           }
           console.log("Created Lyra encoders and decoders.");
 
@@ -71,14 +71,14 @@
               let now = performance.now();
               let encodedAudioDataList = [];
               for (const audioData of audioDataList) {
-                  encodedAudioDataList.push(encoders[version].encode(audioData));
+                  encodedAudioDataList.push(await encoders[version].encode(audioData));
               }
               console.log(`[${version}] Encode Time: ${performance.now() - now} ms`);
               result += `- Encode Time: ${performance.now() - now} ms<br />`;
 
               now = performance.now();
               for (const encodedAudioData of encodedAudioDataList) {
-                  decoders[version].decode(encodedAudioData);
+                  await decoders[version].decode(encodedAudioData);
               }
               console.log(`[${version}] Decode Time: ${performance.now() - now} ms`);
               result += `- Decode Time: ${performance.now() - now} ms<br />`;
