--- examples/lyra_async.js
@@ -1,1 +1,0 @@
-../dist/lyra_async.js
