+++ examples/lyra_sync_worker.js
@@ -1,0 +1,1 @@
+../dist/lyra_sync_worker.js
