+++ src/lyra.ts
@@ -1,6 +1,3 @@
-import loadLyraWasmModule from "./lyra_wasm.js";
-import * as lyra_wasm from "./lyra_wasm.js";
-
 import {
   LYRA_VERSION,
   DEFAULT_SAMPLE_RATE,
@@ -10,303 +7,202 @@ import {
   LyraDecoderOptions,
   LyraEncoderOptions,
   trimLastSlash,
-  checkSampleRate,
-  checkNumberOfChannels,
-  checkBitrate,
 } from "./utils";
 
-const MEMFS_MODEL_PATH = "/tmp/";
-const FRAME_DURATION_MS = 20;
-
-/**
- * Lyra 用の WebAssembly ファイルやモデルファイルを管理するためのクラス
- */
 class LyraModule {
-  private wasmModule: lyra_wasm.LyraWasmModule;
+  private worker: Worker;
 
-  private constructor(wasmModule: lyra_wasm.LyraWasmModule) {
-    this.wasmModule = wasmModule;
+  private constructor(worker: Worker) {
+    this.worker = worker;
   }
 
-  /**
-   * Lyra の WebAssembly ファイルやモデルファイルをロードして {@link LyraModule} のインスタンスを生成する
-   *
-   * @param wasmPath lyra.wasm および lyra.worker.js が配置されているディレクトリのパスないし URL
-   * @param modelPath Lyra 用の *.binarypb および *.tflite が配置されているディレクトリのパスないし URL
-   * @returns 生成された {@link LyraModule} インスタンス
-   */
-  static async load(wasmPath: string, modelPath: string): Promise<LyraModule> {
-    const wasmModule = await loadLyraWasmModule({
-      locateFile: (path) => {
-        return trimLastSlash(wasmPath) + "/" + path;
-      },
-      preRun: (wasmModule) => {
-        const fileNames = ["lyra_config.binarypb", "soundstream_encoder.tflite", "quantizer.tflite", "lyragan.tflite"];
-        for (const fileName of fileNames) {
-          const url = trimLastSlash(modelPath) + "/" + fileName;
-
-          // FIXME(sile): url からのダウンロードに失敗しても catch ができないので fetch で置き換える
-          wasmModule.FS_createPreloadedFile(MEMFS_MODEL_PATH, fileName, url, true, false);
-        }
-      },
+  static load(wasmPath: string, modelPath: string): Promise<LyraModule> {
+    const worker = new Worker(trimLastSlash(wasmPath) + "/lyra_sync_worker.js", {
+      type: "module",
+      name: "lyra_sync_worker",
     });
 
-    return new LyraModule(wasmModule);
+    // モデルは web worker の中でロードされることになるので、
+    // modelPath を事前に絶対 URL に変換して曖昧性がなくなるようにしておく
+    modelPath = new URL(modelPath, document.location.href).toString();
+
+    const promise: Promise<LyraModule> = new Promise((resolve, reject) => {
+      type Response = { data: { type: "LyraModule.load.result"; result: { error?: Error } } };
+      worker.addEventListener(
+        "message",
+        (res: Response) => {
+          const error = res.data.result.error;
+          if (error === undefined) {
+            resolve(new LyraModule(worker));
+          } else {
+            reject(error);
+          }
+        },
+        { once: true }
+      );
+    });
+
+    worker.postMessage({ type: "LyraModule.load", modelPath });
+    return promise;
   }
 
-  /**
-   * {@link LyraEncoder} のインスタンスを生成する
-   *
-   * 生成したインスタンスが不要になったら {@link LyraEncoder.destroy} メソッドを呼び出してリソースを解放すること
-   *
-   * @params options エンコーダに指定するオプション
-   * @returns 生成された {@link LyraEncoder} インスタンス
-   */
-  createEncoder(options: LyraEncoderOptions = {}): LyraEncoder {
-    checkSampleRate(options.sampleRate);
-    checkNumberOfChannels(options.numberOfChannels);
-    checkBitrate(options.bitrate);
+  createEncoder(options: LyraEncoderOptions = {}): Promise<LyraEncoder> {
+    const channel = new MessageChannel();
 
-    const encoder = this.wasmModule.LyraEncoder.create(
-      options.sampleRate || DEFAULT_SAMPLE_RATE,
-      options.numberOfChannels || DEFAULT_CHANNELS,
-      options.bitrate || DEFAULT_BITRATE,
-      options.enableDtx || DEFAULT_ENABLE_DTX,
-      MEMFS_MODEL_PATH
-    );
-    if (encoder === undefined) {
-      throw new Error("failed to create lyra encoder");
-    } else {
-      const frameSize = ((options.sampleRate || DEFAULT_SAMPLE_RATE) * FRAME_DURATION_MS) / 1000;
-      const buffer = this.wasmModule.newAudioData(frameSize);
-      return new LyraEncoder(this.wasmModule, encoder, buffer, options);
-    }
+    const promise: Promise<LyraEncoder> = new Promise((resolve, reject) => {
+      type Response = {
+        data: {
+          type: "LyraModule.createEncoder.result";
+          result: { frameSize: number } | { error: Error };
+        };
+      };
+      channel.port1.addEventListener(
+        "message",
+        (res: Response) => {
+          const result = res.data.result;
+          if ("error" in result) {
+            reject(result.error);
+          } else {
+            resolve(new LyraEncoder(channel.port1, result.frameSize, options));
+          }
+        },
+        { once: true }
+      );
+      channel.port1.start();
+    });
+
+    this.worker.postMessage({ type: "LyraModule.createEncoder", port: channel.port2, options }, [channel.port2]);
+    return promise;
   }
 
-  /**
-   * {@link LyraDecoder} のインスタンスを生成する
-   *
-   * 生成したインスタンスが不要になったら {@link LyraDecoder.destroy} メソッドを呼び出してリソースを解放すること
-   *
-   * @params options デコーダに指定するオプション
-   * @returns 生成された {@link LyraDecoder} インスタンス
-   */
-  createDecoder(options: LyraDecoderOptions = {}): LyraDecoder {
-    checkSampleRate(options.sampleRate);
-    checkNumberOfChannels(options.numberOfChannels);
+  createDecoder(options: LyraDecoderOptions = {}): Promise<LyraDecoder> {
+    const channel = new MessageChannel();
 
-    const decoder = this.wasmModule.LyraDecoder.create(
-      options.sampleRate || DEFAULT_SAMPLE_RATE,
-      options.numberOfChannels || DEFAULT_CHANNELS,
-      MEMFS_MODEL_PATH
-    );
-    if (decoder === undefined) {
-      throw new Error("failed to create lyra decoder");
-    } else {
-      const buffer = this.wasmModule.newBytes();
-      return new LyraDecoder(this.wasmModule, decoder, buffer, options);
-    }
+    const promise: Promise<LyraDecoder> = new Promise((resolve, reject) => {
+      type Response = {
+        data: {
+          type: "LyraModule.createDecoder.result";
+          result: { frameSize: number } | { error: Error };
+        };
+      };
+      channel.port1.addEventListener(
+        "message",
+        (res: Response) => {
+          const result = res.data.result;
+          if ("error" in result) {
+            reject(result.error);
+          } else {
+            resolve(new LyraDecoder(channel.port1, result.frameSize, options));
+          }
+        },
+        { once: true }
+      );
+      channel.port1.start();
+    });
+
+    this.worker.postMessage({ type: "LyraModule.createDecoder", port: channel.port2, options }, [channel.port2]);
+    return promise;
   }
 }
 
-/**
- * Lyra のエンコーダ
- */
 class LyraEncoder {
-  private wasmModule: lyra_wasm.LyraWasmModule;
-  private encoder: lyra_wasm.LyraWasmEncoder;
-  private buffer: lyra_wasm.AudioData;
-
-  /**
-   * 現在のサンププリングレート
-   */
+  readonly port: MessagePort;
   readonly sampleRate: number;
-
-  /**
-   * 現在のチャネル数
-   */
   readonly numberOfChannels: number;
-
-  /**
-   * 現在のエンコードビットレート
-   */
   readonly bitrate: number;
-
-  /**
-   * DTX が有効になっているかどうか
-   */
   readonly enableDtx: boolean;
-
-  /**
-   * 一つのフレーム（{@link LyraEncoder.encode} メソッドに渡す音声データ）に含めるサンプル数
-   */
   readonly frameSize: number;
 
-  /**
-   * @internal
-   */
-  constructor(
-    wasmModule: lyra_wasm.LyraWasmModule,
-    encoder: lyra_wasm.LyraWasmEncoder,
-    buffer: lyra_wasm.AudioData,
-    options: LyraEncoderOptions
-  ) {
-    this.wasmModule = wasmModule;
-    this.encoder = encoder;
-    this.buffer = buffer;
-
+  constructor(port: MessagePort, frameSize: number, options: LyraEncoderOptions) {
+    this.port = port;
+    this.frameSize = frameSize;
     this.sampleRate = options.sampleRate || DEFAULT_SAMPLE_RATE;
     this.numberOfChannels = options.numberOfChannels || DEFAULT_CHANNELS;
     this.bitrate = options.bitrate || DEFAULT_BITRATE;
     this.enableDtx = options.enableDtx || DEFAULT_ENABLE_DTX;
-    this.frameSize = buffer.size();
   }
 
-  /**
-   * 20ms 分の音声データをエンコードする
-   *
-   * @params audioData エンコード対象の音声データ
-   * @returns エンコード後のバイト列。もし DTX が有効で音声データが無音な場合には undefined が代わりに返される。
-   *
-   * @throws
-   *
-   * 以下のいずれかに該当する場合には例外が送出される:
-   * - 入力音声データが 20ms 単位（サンプル数としては {@link LyraEncoder.frameSize}）ではない
-   * - その他、何らかの理由でエンコードに失敗した場合
-   */
-  encode(audioData: Int16Array): Uint8Array | undefined {
-    if (audioData.length !== this.frameSize) {
-      throw new Error(
-        `expected an audio data with ${this.frameSize} samples, but got one with ${audioData.length} samples`
+  // TODO: audioData が transfer されることを書く
+  encode(audioData: Int16Array): Promise<Uint8Array | undefined> {
+    const promise: Promise<Uint8Array | undefined> = new Promise((resolve, reject) => {
+      type Response = {
+        data: {
+          type: "LyraEncoder.encode.result";
+          result: { encodedAudioData: Uint8Array | undefined } | { error: Error };
+        };
+      };
+      this.port.addEventListener(
+        "message",
+        (res: Response) => {
+          const result = res.data.result;
+          if ("error" in result) {
+            reject(result.error);
+          } else {
+            resolve(result.encodedAudioData);
+          }
+        },
+        { once: true }
       );
-    }
+    });
 
-    this.wasmModule.copyInt16ArrayToAudioData(this.buffer, audioData);
-
-    const result = this.encoder.encode(this.buffer);
-
-    if (result === undefined) {
-      throw new Error("failed to encode");
-    } else {
-      try {
-        const encodedAudioData = new Uint8Array(result.size());
-        for (let i = 0; i < encodedAudioData.length; i++) {
-          encodedAudioData[i] = result.get(i);
-        }
-
-        if (encodedAudioData.length === 0) {
-          // DTX が有効、かつ、 audioData が無音ないしノイズだけを含んでいる場合にはここに来る
-          return undefined;
-        }
-        return encodedAudioData;
-      } finally {
-        result.delete();
-      }
-    }
+    this.port.postMessage({ type: "LyraEncoder.encode", audioData, bitrate: this.bitrate }, [audioData.buffer]);
+    return promise;
   }
 
-  /**
-   * エンコードビットレートを変更する
-   *
-   * @params bitrate 変更後のビットレート
-   */
-  setBitrate(bitrate: number): void {
-    checkBitrate(bitrate);
-
-    if (!this.encoder.setBitrate(bitrate)) {
-      throw new Error(`failed to update bitrate from ${this.bitrate} to ${bitrate}`);
-    }
-  }
-
-  /**
-   * エンコーダ用に確保したリソースを解放する
-   */
   destroy(): void {
-    this.encoder.delete();
-    this.buffer.delete();
+    this.port.postMessage({ type: "LyraEncoder.destroy" });
+    this.port.close();
   }
 }
 
-/**
- * Lyra のデコーダ
- */
 class LyraDecoder {
-  private wasmModule: lyra_wasm.LyraWasmModule;
-  private decoder: lyra_wasm.LyraWasmDecoder;
-  private buffer: lyra_wasm.Bytes;
-
-  /**
-   * 現在のサンププリングレート
-   */
+  readonly port: MessagePort;
   readonly sampleRate: number;
-
-  /**
-   * 現在のチャネル数
-   */
   readonly numberOfChannels: number;
-
-  /**
-   * 一つのフレーム（{@link LyraEncoder.decode} メソッドの返り値の音声データ）に含まれるサンプル数
-   */
   readonly frameSize: number;
 
-  /**
-   * @internal
-   */
-  constructor(
-    wasmModule: lyra_wasm.LyraWasmModule,
-    decoder: lyra_wasm.LyraWasmDecoder,
-    buffer: lyra_wasm.Bytes,
-    options: LyraDecoderOptions
-  ) {
-    this.wasmModule = wasmModule;
-    this.decoder = decoder;
-    this.buffer = buffer;
-
+  constructor(port: MessagePort, frameSize: number, options: LyraDecoderOptions) {
+    this.port = port;
+    this.frameSize = frameSize;
     this.sampleRate = options.sampleRate || DEFAULT_SAMPLE_RATE;
     this.numberOfChannels = options.numberOfChannels || DEFAULT_CHANNELS;
-
-    this.frameSize = (this.sampleRate * FRAME_DURATION_MS) / 1000;
   }
 
-  /**
-   * {@link LyraEncoder.encode} メソッドによってエンコードされた音声データをデコードする
-   *
-   * @params encodedAudioData デコード対象のバイナリ列ないし undefined
-   * @returns デコードされた 20ms 分の音声データ。undefined が渡された場合には代わりにコンフォートノイズが生成される。
-   */
-  decode(encodedAudioData: Uint8Array | undefined): Int16Array {
-    if (encodedAudioData !== undefined) {
-      this.buffer.resize(0, 0); // clear() を使うと「関数が存在しない」というエラーが出るので resize() で代用
-      for (const v of encodedAudioData) {
-        this.buffer.push_back(v);
-      }
-      if (!this.decoder.setEncodedPacket(this.buffer)) {
-        throw new Error("failed to set encoded packet");
-      }
-    }
+  // TODO: encodedAudioData が transfer されることを書く
+  decode(encodedAudioData: Uint8Array | undefined): Promise<Int16Array> {
+    const promise: Promise<Int16Array> = new Promise((resolve, reject) => {
+      type Response = {
+        data: {
+          type: "LyraDecoder.decode.result";
+          result: { audioData: Int16Array } | { error: Error };
+        };
+      };
+      this.port.addEventListener(
+        "message",
+        (res: Response) => {
+          const result = res.data.result;
+          if ("error" in result) {
+            reject(result.error);
+          } else {
+            resolve(result.audioData);
+          }
+        },
+        { once: true }
+      );
+    });
 
-    const result = this.decoder.decodeSamples(this.frameSize);
-
-    if (result === undefined) {
-      throw Error("failed to decode samples");
+    if (encodedAudioData === undefined) {
+      this.port.postMessage({ type: "LyraDecoder.decode", encodedAudioData });
+    } else {
+      this.port.postMessage({ type: "LyraDecoder.decode", encodedAudioData }, [encodedAudioData.buffer]);
     }
-    try {
-      const audioData = new Int16Array(this.frameSize);
-      this.wasmModule.copyAudioDataToInt16Array(audioData, result);
-      return audioData;
-    } finally {
-      result.delete();
-    }
+    return promise;
   }
 
-  /**
-   * デコーダ用に確保したリソースを解放する
-   */
   destroy(): void {
-    this.decoder.delete();
-    this.buffer.delete();
+    this.port.postMessage({ type: "LyraDecoder.destroy" });
+    this.port.close();
   }
 }
 
-export { LyraModule, LyraDecoder, LyraEncoder, LyraEncoderOptions, LyraDecoderOptions, LYRA_VERSION };
+export { LYRA_VERSION, LyraDecoderOptions, LyraEncoderOptions, LyraModule, LyraEncoder, LyraDecoder };
